Added navbar + color edits
Added navbar, as well as an additional row in the grid to accommodate it.

Changed some of the rgb values.

Added a footer so the bottom wouldn't cut off.

Took away universal selector bc it was adding all this filler text but now my padding is a little wonky. Such is life.