Modified palette in CSS
Modified the palette out of grayscale into color.
Currently, the main accent color is a pale desaturated yellow, the majority of the page is green-blue tinted gray, and other parts are untinted gray values.
I have to rework the images, font, and layout before I decide if or not I like it.